Add viewModel
