+++ Stopwatch/app/src/main/java/com/gb/stopwatch/State.kt
@@ -1,0 +1,12 @@
+package com.gb.stopwatch
+
+sealed class State {
+    data class Paused(
+        val elapsedTime: Long
+    ) : State()
+
+    data class Running(
+        val startTime: Long,
+        val elapsedTime: Long
+    ) : State()
+}
