+++ Stopwatch/app/src/main/java/com/gb/stopwatch/view/MainActivity.kt
@@ -1,0 +1,39 @@
+package com.gb.stopwatch.view
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.ViewModelProvider
+import com.gb.stopwatch.databinding.ActivityMainBinding
+
+class MainActivity : AppCompatActivity() {
+
+    private lateinit var binding: ActivityMainBinding
+
+    private val viewModel: MainActivityViewModel by lazy {
+        ViewModelProvider(this).get(MainActivityViewModel::class.java)
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        viewModel.liveData.observe(this) {
+            binding.textTime.text = it
+        }
+
+        viewModel.collect()
+
+        binding.buttonStart.setOnClickListener {
+            viewModel.start()
+        }
+        binding.buttonPause.setOnClickListener {
+            viewModel.pause()
+        }
+        binding.buttonStop.setOnClickListener {
+            viewModel.stop()
+        }
+    }
+}
+
+
