+++ Stopwatch/app/src/main/java/com/gb/stopwatch/view/MainActivityViewModel.kt
@@ -1,0 +1,54 @@
+package com.gb.stopwatch.view
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import com.gb.stopwatch.*
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.SupervisorJob
+import kotlinx.coroutines.flow.collect
+import kotlinx.coroutines.launch
+
+class MainActivityViewModel : ViewModel() {
+
+    private val mutableLiveData: MutableLiveData<String> = MutableLiveData()
+    val liveData: LiveData<String> = mutableLiveData
+
+    private val timestampProvider = object : TimestampProvider {
+        override fun getMilliseconds(): Long {
+            return System.currentTimeMillis()
+        }
+    }
+
+    private val stopwatchListOrchestrator = StopwatchListOrchestrator(
+        StopwatchStateHolder(
+            StopwatchStateCalculator(
+                timestampProvider,
+                ElapsedTimeCalculator(timestampProvider)
+            ),
+            ElapsedTimeCalculator(timestampProvider),
+            TimestampMillisecondsFormatter()
+        ), CoroutineScope(Dispatchers.Main + SupervisorJob())
+    )
+
+    fun collect() {
+        CoroutineScope(Dispatchers.Main + SupervisorJob()).launch {
+            stopwatchListOrchestrator.ticker.collect {
+                mutableLiveData.value = it
+            }
+        }
+    }
+
+    fun start() {
+        stopwatchListOrchestrator.start()
+    }
+
+    fun pause() {
+        stopwatchListOrchestrator.pause()
+    }
+
+    fun stop() {
+        stopwatchListOrchestrator.stop()
+    }
+}
